Replaces all smart quotes with straight quotes.
Although some would argue that smart quotes are the right choice for general
typography, mixing smart and dumb quotes is inconsistent, and the HTML document
already needs straight quotes for string literals.

It's reasonable to decide that text-quotes should be smart and code-quotes
should be straight, but if so, it's an easy change, and I'm opting for
consistency in the absence of strong opinions.