Adjusts whitespace and line lengths.
It appears that the line limit for this file is 80 characters. However, with
such a paragraph-heavy text page, it may be preferable to remove or increase
the line limit (and turn on autowrap) so that editing the text later doesn't
conflict with the line length.

Formatted using https://www.freeformatter.com/html-formatter.html with some
modifications.